Use MVP concept
