+++ js/game/game.js
@@ -1,209 +1,99 @@
 
-import * as utils from './../utils';
-
-import levels from './data-levels';
-
-import ingameLevel from '../level/level';
-import greetingScreen from '../greeting/greeting';
-import statsScreen from '../stats/stats';
-
-
-const extraPoints = {
-  fast: `Бонус за скорость:`,
-  heart: `Бонус за жизни:`,
-  slow: `Штраф за медлительность:`
-};
-
-let levelTimer = null;
+import {renderScreen} from '../data/data';
+import {loadLevels} from '../data/data';
+import {getLevelResult} from '../data/data';
+import {state as initState} from '../data/data';
+import {rules} from '../data/data';
+import {types} from '../data/data';
+import {levels} from '../data/data-levels';
+import GameView from './game-view';
+import Application from '../application';
 
 
-export const rules = Object.freeze({
-  levelTime: 30,
-  slowTime: 20,
-  quickTime: 10,
-  points: Object.freeze({
-    correct: 100,
-    fast: 50,
-    slow: -50,
-    wrong: 0,
-    unknown: 0,
-    heart: 50
-  }),
-  maxLives: 3,
-  levelsCount: levels.length
-});
+class GamePresenter {
+  constructor(state) {
+    //loadLevels(levels, () => {});
 
-export const state = Object.freeze({
-  level: 0,
-  lives: rules.maxLives,
-  name: `Unknown`,
-  results: Object.freeze(new Array(rules.levelsCount).fill(`unknown`))
-});
+    this.state = state;
+    this.level = levels[state.level];
 
-export function renderScreen(screen) {
+    this.view = new GameView(this.state, Object.assign({},
+      this.level,
+      types[this.level.type]
+    ));
 
-  const viewport = document.getElementById(`main`);
+    this.gameTimer = null;
+  }
 
-  viewport.innerHTML = ``;
-  viewport.appendChild(screen.element);
-}
+  _startGame() {
 
-export function loadLevels(onLoadCompleted) {
+    const TIMER_DELAY = 1000;
 
-  let count = levels.length;
+    let timerTiks = rules.levelTime;
 
-  levels.forEach((level, index) => {
+    this.gameTimer = setInterval(() => {
 
-    utils.loadImages(level.src, (imgs) => {
+      this.view.levelTime = timerTiks--;
 
-      levels[index].img = imgs;
-      count--;
-
-      // console.log(`Imgs of level ${index} is loaded, ${count} left`);
-
-      if (!count && typeof onLoadCompleted === `function`) {
-        onLoadCompleted();
+      if (!timerTiks) {
+        this._endGame(this.state);
       }
-    });
-  });
-}
-
-export function renderLevel(curState) {
-
-  const levelScreen = ingameLevel(curState, levels[curState.level]);
-
-  renderScreen(levelScreen);
-
-  startLevel(curState, (timerTiks) => {
-    levelScreen.levelTime = timerTiks;
-  });
-}
-
-export function renderNextLevel(curState) {
-
-  if ((curState.lives >= 0) && (curState.level + 1) < rules.levelsCount) {
-    renderLevel(Object.assign({}, curState, {
-      level: curState.level + 1
-    }));
-  } else {
-    renderScreen(statsScreen(curState));
+    }, TIMER_DELAY);
   }
-}
 
-export function getLevelResult(levelTime, levelPassed) {
+  _endGame(curState, levelTime = 0, levelPassed = false) {
 
-  if (!levelPassed || levelTime <= 0) {
-    return `wrong`;
-  } else if (levelPassed && levelTime < rules.quickTime) {
-    return `fast`;
-  } else if (levelPassed && levelTime > rules.slowTime) {
-    return `slow`;
-  } else if (levelPassed) {
-    return `correct`;
-  } else {
-    return `wrong`;
-  }
-}
+    clearInterval(this.gameTimer);
 
-export function startLevel(curState, onLevelTime) {
+    const result = getLevelResult(levelTime, levelPassed);
 
-  const TIMER_DELAY = 1000;
-
-  let timerTiks = rules.levelTime;
-
-  levelTimer = setInterval(() => {
-
-    timerTiks--;
-
-    if (typeof onLevelTime === `function`) {
-      onLevelTime(timerTiks);
-    }
-
-    if (!timerTiks) {
-      finishLevel(curState);
-    }
-
-  }, TIMER_DELAY);
-}
-
-export function finishLevel(curState, levelTime, levelPassed) {
-
-  clearInterval(levelTimer);
-
-  const result = getLevelResult(levelTime, levelPassed);
-
-  const newState = Object.assign({}, curState, {
-    lives: (result === `wrong`)
-      ? curState.lives - 1
-      : curState.lives,
-    results: curState.results.slice()
-  });
-
-  newState.results[curState.level] = result;
-
-  renderNextLevel(newState);
-}
-
-export function countResults(results, value) {
-  return results.filter((result) => result === value).length;
-}
-
-export function getLivesCount(results) {
-
-  const lives = rules.maxLives - countResults(results, `wrong`);
-
-  return (lives >= 0) ? lives : 0;
-}
-
-export function getPoints(results) {
-  return results.filter((result) => {
-    return Math.abs(rules.points[result]);
-  }).length * rules.points.correct;
-}
-
-export function getTotalPoints(results) {
-  return getPoints(results) + Object.keys(extraPoints).map((key) => {
-
-    const keyCount = (key === `heart`)
-      ? getLivesCount(results)
-      : countResults(results, key);
-
-    return keyCount * rules.points[key];
-
-  }).reduce((pValue, cValue) => pValue + cValue);
-}
-
-export function getExtraPointsList(results) {
-
-  return Object.keys(extraPoints).map((key) => {
-
-    const keyCount = (key === `heart`)
-      ? getLivesCount(results)
-      : countResults(results, key);
-
-    return {
-      name: key,
-      title: extraPoints[key],
-      count: keyCount,
-      points: Math.abs(rules.points[key]),
-      totalPoints: keyCount * rules.points[key]
-    };
-  });
-}
-
-export function start(curState, userName) {
-
-  loadLevels(() => {
-    renderLevel(Object.assign({}, curState, {
-      name: userName,
+    const newState = Object.assign({}, curState, {
+      lives: (result === `wrong`)
+        ? curState.lives - 1
+        : curState.lives,
       results: curState.results.slice()
-    }));
-  });
+    });
+
+    newState.results[curState.level] = result;
+
+    this._nextGame(newState);
+  }
+
+  _nextGame(curState) {
+
+    if ((curState.lives >= 0) && (curState.level + 1) < rules.levelsCount) {
+
+      this.state = Object.assign({}, curState, {
+        level: curState.level + 1
+      });
+
+      this.level = levels[this.state.level];
+
+      this.view = new GameView(this.state, Object.assign({},
+        this.level,
+        types[this.level.type]
+      ));
+
+      this.init();
+
+    } else {
+      Application.showStats(curState.results);
+    }
+  }
+
+  init() {
+
+    renderScreen(this.view);
+
+    this.view.onLevelFinished = (curState, levelTime, isAnswerRight) => {
+      this._endGame(curState, levelTime, isAnswerRight);
+    };
+
+    this.view.onBackButtonClick = () => {
+      clearInterval(this.gameTimer);
+      Application.showGreeting();
+    };
+  }
 }
 
-export function reset() {
-
-  clearInterval(levelTimer);
-
-  renderScreen(greetingScreen());
-}
+export default new GamePresenter(initState);
