+++ js/rules/rules.js
@@ -1,19 +1,36 @@
 
-import * as game from '../game/game';
+import {renderScreen} from '../data/data';
+import {state} from '../data/data';
 import RulesView from './rules-view';
+import Application from '../application';
+import {loadLevels} from '../data/data-levels';
+import {getLevel} from '../data/data-levels';
 
+class RulesPresenter {
+  constructor() {
+    this.view = new RulesView();
+  }
 
-export default () => {
+  init() {
 
-  const rulesScreen = new RulesView();
+    renderScreen(this.view);
 
-  rulesScreen.onContinueButtonClick = (userName) => {
-    game.start(game.state, userName);
-  };
+    this.view.onContinueButtonClick = (userName) => {
 
-  rulesScreen.onBackButtonClick = () => {
-    game.reset();
-  };
+      loadLevels((levels) => {
 
-  return rulesScreen;
-};
+        Application.showGame(Object.assign({}, state, {
+          name: userName,
+          levels: levels,
+          results: state.results.slice()
+        }));
+      });
+    };
+
+    this.view.onBackButtonClick = () => {
+      Application.showGreeting();
+    };;
+  }
+}
+
+export default new RulesPresenter();
